Pass testData.task_count_by_state to TestFactory::create() (#844)
