+++ src/Controller/View/Test/ProgressController.php
@@ -191,6 +191,7 @@ class ProgressController extends AbstractBaseViewController
             'amendments' => $remoteTest->getAmmendments(),
             'crawl' => $remoteTest->getCrawl(),
             'task_types' => $remoteTest->getTaskTypes(),
+            'task_count_by_state' => $remoteTest->getRawTaskCountByState(),
         ]);
         $decoratedTest = new DecoratedTest($testModel);
