+++ src/Controller/View/Test/Results/ByTaskTypeController.php
@@ -213,6 +213,7 @@ class ByTaskTypeController extends AbstractBaseViewController
             'amendments' => $remoteTest->getAmmendments(),
             'crawl' => $remoteTest->getCrawl(),
             'task_types' => $remoteTest->getTaskTypes(),
+            'task_count_by_state' => $remoteTest->getRawTaskCountByState(),
         ]);
         $decoratedTest = new DecoratedTest($testModel);
