+++ src/Controller/View/Test/Results/ResultsController.php
@@ -227,6 +227,7 @@ class ResultsController extends AbstractBaseViewController
             'amendments' => $remoteTest->getAmmendments(),
             'crawl' => $remoteTest->getCrawl(),
             'task_types' => $remoteTest->getTaskTypes(),
+            'task_count_by_state' => $remoteTest->getRawTaskCountByState(),
         ]);
         $decoratedTest = new DecoratedTest($testModel);
