+++ tests/Unit/Model/DecoratedTestListTest.php
@@ -10,6 +10,7 @@ use App\Model\RemoteTest\RemoteTest;
 use App\Model\Test\DecoratedTest;
 use App\Services\TestCompletionPercentCalculator;
 use App\Services\TestFactory;
+use App\Services\TestTaskCountByStateNormaliser;
 use App\Tests\Services\ObjectReflector;
 
 class DecoratedTestListTest extends \PHPUnit\Framework\TestCase
@@ -34,7 +35,10 @@ class DecoratedTestListTest extends \PHPUnit\Framework\TestCase
         $entity = TestEntity::create($testId);
         $remoteTest = new RemoteTest([]);
 
-        $testFactory = new TestFactory(new TestCompletionPercentCalculator());
+        $testFactory = new TestFactory(
+            new TestCompletionPercentCalculator(),
+            new TestTaskCountByStateNormaliser()
+        );
         $test = $testFactory->create($entity, $remoteTest, []);
 
         $decoratedTest = new DecoratedTest($test);
